afinter: current_internal_source was used for 2 things:
1) Is there any configured internal source in the config
2) Does the configured internal source free to send or not

If there isn't configured internal source,
the messages shouldn't be stored in the internal queue,
because there is no one to read them.

But if there is a configured internal source,
the messages should be stored in the internal queue,
independently from the queue source is ready to work or not.

There is a side effect of this patch, that if there was no
internal source, before reload, but after reload it exists,
then the internal messages generated before reload will be lost

Fixes: #1494

Signed-off-by: Juhász Viktor <juhasz.viktor81@gmail.com>