afsocket: Remove invalid assertion from afsocket_dd_restore_writer.
Asserting that self->writer is null seems to be incorrect.

If the config being initialized is new, self->writer is guaranteed
to be null, however if we are restoring an old config after initializing
a new configuration failed, an existing writer may be present if keep-alive
is off (old_config self->writer is set to null if the connection is saved)

Instead check that self->writer is non NULL before attempting to restore one.
If an existing writer is present, the stored writer should be freed by
_reload_store_item_free().

Fixes #1722

Signed-off-by: Sam Stephenson <sam.stephenson@alliedtelesis.co.nz>