filters: fix filter evaluation in case of contexts with multiple elements
Filter expressions support contexts, that is, they can be applied to
a list of LogMessage instances instead of just one.

The reasons for this functionality are:
  * filter expressions may have template strings in them
  * template strings may use the context information
  * without filters supporting a context based API, embedded templates
    would not get the context

E.g. if you wanted to check on the $(context-length) within a conditional
expression in db-parser(), you'll need the context to be propagated
through the filter.

There's one case up for interpretation: the primitive filter
expressions (e.g. match), would need to be applied only to
one LogMessage instance.

And this was buggy, earlier the _first_ message was evaluated, whereas
template strings always work on the _last_ one.

Also, the last one is the "current" message in db-parser and grouping-by,
so clearly this is incorrect behavior.

Unfortunately _none_ of the unit tests covered this case so far, so even
with this non-trivial change, tests passed properly.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>