Emit less confusing warnings for diskq
Using the word "omitted" may be confusing, as it sounds like something
should have been there, but was not taken into account. Let's use
"ignored" instead, as for something that *should not* be there in the
first place (and therefore was ignored).
Also be more verbose and suggest the other configuration parameter.

Fixes #1733.

Signed-off-by: Jakub Jankowski <shasta@toxcorp.com>