kvformat: fix $(format-welf) space printing logic
$(format-welf) puts spaces between name-value pairs, and it reused
result->len to track whether anything has already been added to the string
to determine if a space needs to be added or not.

However, should anything be in the string already, we added another space,
rendering incorrect behaviour in cases where $(format-welf) had a prefix.

e.g.

"foo$(format-welf *)bar"

Incorrectly had a space between "foo" and the first name-value pair.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>