dbld: integrate packaging/debian and packaging/rhel into the build
Earlier, the docker based build system in dbld/ was lacking a key
component: the debian/ directory and the spec file needed to build
RPMs.

In the last weeks those were added into packaging/ subdirectory, but
it was not integrated into this flow.

This patch adds just that, and from now on allows the creation of
  - distribution tarballs
  - deb
  - rpm packages

in both snapshot and release mode.

To create a snapshot Debian package:

$ dbld/rules deb

To create a release Debian package:

$ dbld/rules MODE=release deb

In snapshot mode, it automatically generates a time based version number,
and updates that into the debian packaging automatically.

RPM compilation works the same way, you just use "rpm" in place of "deb".

In release mode, it expects the packaging directories to be set up properly,
e.g. they should be updated to the latest upstream source. The
scripts validate that this is the case.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>