kv-scanner: simplify class with the assumption that all call sites would be using a local variable
This patch gets rid off the reuse of KVScanner instances accross different
messages in production.  KVScanners have never been threadsafe, and thus a
cloning of it was introduced for every message received.

This patch gets rid off the clone and rather takes the performance hit of
having to allocate a couple of GString instances, and making the KVScanner
instance a local variable in the process function.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>