afprog: Do not restart the child on command-not-found
The problem this fixes is that when the destination tries to use a program that
does not exist, we end up trying to restart it endlessly, forever, without
logging anything about it on a reasonable log level. This spams the logs, and is
very unhelpful anyway. Instead, when encountering a command-not-found error, we
simply stop restarting, and stop the destination instead.

As an added bonus, when restarting the child, we now log that at the info level,
not just on verbose.

Signed-off-by: Gergely Nagy <algernon@balabit.com>