grammar: Make sure that nonnegative and positive integers fit into 32 bits
Every time `nonnegative_integer` or `positive_integer` is used in the grammar,
the numbers end up being stored on 32 bits (except for `disk-buf-size`, which is
64-bit wide). Not properly bounds-checking these numbers can result in us ending
up with a negative value.

Seeing as there is only one exception, this patch changes `nonnegative_integer`
and `positive_integer` to do proper bounds checking too, and error out if the
value does not fit. For `disk-buf-size`, we introduce `nonnegative_integer64`,
local to that part of the grammar.

Fixes #1820.

Signed-off-by: Gergely Nagy <algernon@balabit.com>