grammar: Rearrange nonnegative_integer{,64}
Instead of having `nonnegative_integer64` local to diskq's grammar, move it into
`lib/cfg-grammar.y`. When there, make `nonnegative_integer` a derivative of
it (and do the same for `positive_integer` and `positive_integer64` for the sake
of symmetry).

The error message has been slightly improved too in the process, and redundancy
removed (we don't need to check for `>= G_MININT32`, the previous checks already
imply this).

Signed-off-by: Gergely Nagy <algernon@balabit.com>