add-contextual-data: option to filter messages
To better control to which log messages you add contextual data, you can use filters as selectors. In this case, the first column of the CSV database file must contain the name of a filter. For each message, syslog-ng evaluates the filters in the order they appear in the database file. If a filter matches the message, syslog-ng adds the name-value pair related to the filter.

Signed-off-by: Andras Mitzki <mitzkia@gmail.com>
Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>
Signed-off-by: László Várady <laszlo.varady@balabit.com>
Signed-off-by: Máté Farkas <mate.farkas@balabit.com>
Signed-off-by: Viktor Juhasz <viktor.juhasz@balabit.com>
Signed-off-by: kokan <peter.kokai@balabit.com>