NVHandleDescriptors: adding NVHandleDescArray implementation
A new array implementation added for storing NVHandleDesc with the
purpose of replacing the GArray in nvtable in the next patch.

As with the original behaviour: elements can only be added, adding
must happen under lock. The difference is that reading must not happen
under lock. When the array needs to be expanded, instead of realloc, a
completely new memory area is allocated, the original data is copied
to the new area. But the original area is not freed until deinit.
This means the references in the running jobs remain valid. So there
is no need for locking while reading. As only append is called to
the array, the size of the reserved but not used memory is bounded by
the current size of the array: sum of two powers below a two power is
the two power minus one.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>