elastic-v2: HttpBulkMessageProcessor: fix a message drop issue
Each time when the flush() (which is called by send()) returns with
false, we increment a counter.
When the value of this counter reach a limit (max retries), syslog-ng
drops the messages.
Unfortunately this mechanism is not working for elastic http bulk
because the counter never reach the limit.
The reason of this is that we call init() in ESHttpClient.open() after
the java dst is suspended, which is happen when an error is reported by
send, and the init creates a new Bulk.Builder and reset messagecounter.
The next send will be succeeded as the messagecounter has been reseted
(send() returning true until the batch size has not been reached, then
it calls flush which fails again, java dst suspended again, then open is
called which calls init, and it will happen again and again, until all
the messages are dropped silently).

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>