cfg: cfg_new_snippet not requires license module
The license force load was hard coded into the lexer, if configured via build time.
This modification removes the license from lexer, and behavies differently for cfg_new and cfg_new_snippet.
The cfg_new (if configured durring build) forces to preload the license.
The cfg_new_snippet does not force to load the license module.

Signed-off-by: kokan <peter.kokai@balabit.com>