afsocket-source: support asynchronous transport mapper init
In case of stream sockets it is possible that a resource become
available only at runtime.
Until that point syslog-ng won't accept connections from clients.
So init can be split into two phase:
1) socket creation, bind
2) listen and accept

Currently we have only one use case: password protected SSL keys.

Other solution could be to move back the initialization of the TLS context
to connection time. The problem with this approach is that until
syslog-ng does not have the password, a client can establish connection
to the server.

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>