cfg-grammar: add implementation for if {} elif {}  else {} blocks
We are already able to do conditional evaluation with a bit arcane syntax:

    junction {
      channel {
        filter { EXPR1 };
        ...
        flags(final);
      };
      channel {
        filter { EXPR2 };
        ...
        flags(final);
      };
      channel {
        flags(final);
      };
    };

However, this is not very intuitive. This patch adds syntactic sugar using
the if / else keywords to achieve the same, such as:

    if (message('foo')) {
       ...
    } elif (message('bar')) {
       ...
    } else {
       ...
    }

We actually have two formats of the conditional expressions:

1) the one with an explicit filter expression

    if (message('foo')) {
       parser { date-parser(); };
    } else {
        ...
    };

2) the one where the condition is embedded in the log path

    if {
        filter { message('foo')); };
        parser { date-parser(); };
    } else {
        ...
    };

The difference between the two is that the second format would consider all
filters and all parsers as the condition, combined.  The first one only uses the
filter expression in if().

The difference in evaluation of the two formats are:

  1) if the message contains 'foo' and the date-parser() fails, the message
     is dropped and execution does not continue along the else path.  If if
     does not contain 'foo' the else is obviously taken.

  2) if the message contains 'foo' and the date-parser() fails the else
     branch is taken, just as if the message does not contain 'foo'.

This change should make the configuration for complex of stuff a lot
simpler.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>