diskq: write serialized messages in chunks in qdisk save queue
In the past qdisk_save_queue only wrote the serialized data to the diskq
when all messages have been popped from the queue and serialized.
During serialization each serialized message is appended to the
same string buffer which could lead to huge memory usage (proportional to
the log-fifo-size).

With this patch we define an internal memory limit of the string buffer.
When the limit is reached, we write the serialized messages to the
disk-buffer and truncate the buffer.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>