Moved formatted_key from KVParser into a local variable in _process.
KVScanner is always initialized on the heap to avoid race conditions,
so the kv_scanner_new function was only used from the tests.
Eliminated the use of new/free from the test_linux_audit_scanner, and
made a local (proper) version of create_kv_scanner for the
test_kv_scanner testcase.