diskq: Fix initialization of QDisk object when restarting disk-queue.
As QDisk objects can be reused if a LogQueueDisk child-class creates a
new disk-queue file (known as restart), QDisk object should be properly
initialized before the QDisk object would be first used (in qdisk_start.

i.e. init_instance functionality is missing.
For this the `file_size` member is initialized in qdisk_init and qdisk_start.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>