python: move string related helpers from compat to python-helpers
These functions are more alike than different, so unify their behavior
for both Python 2 & 3, that is accept both bytes and unicode objects
in both versions.

This still requires some version dependent code due to a missing
function in the Python 2 API, but I've felt that it is easier to read
it this way, than duplicating the logic in two separate modules.

The patch also migrates all code to use the new functions, instead of
open-coded similar checks.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>