logmsg: placeholder for procesesd timestamps
The C_ prefixed timestamps always evaluated as the actual time. This gives the possibility to get the time when the destination handles the message (or any LogPipe expands macro), as there could be a difference between the received and processed time. But this timestamp is not stored into the internal LogMessage structure, in case of the message is serialized and deserialized this information is lost.

The serialized message has a space for one timestamp. This reuses that space to store a new timestamp PROCESSED. That timestamp should contains the time when the destination handles the LogMessage (start processing), but it should be updated before macro expansion.

IMPORTANT
The PROCESSED timestamp is not updated for the destination, that is something the destination should take care of if it has to store that time as well.

This timestamp also exposed as macro the same way as received, sent timestamp related macros (see R_, S_).
The corresponding prefix is P_. (For example: P_UNIXTIME.)

Signed-off-by: kokan <peter.kokai@balabit.com>