FileReader: refactor - add NC_FILE_MISSING notification
In case of follow file mode, when the file is deleted, this patch will
emit an NC_FILE_MISSING notification. When such scenario is detected,
the code is rescheduled the follow timer. This reschedule is moved
behind a callback that can be overwritten in the next patch.

The purpose is FileReader users can manage the lifecycle of FileReader
with the missing notification, specially: wildcard file reader can
deinit the filereader.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>