Wildcard: deinitializing file reader when removed with notification
The original verion of wildcard filesource deinitialized the the
FileReader in the main thread, but at the same time message handling
could happen in worker threads. As a result a crash could happen,
because a message could be inserted to a deinitialized pipe.

This patch uses the FileReader->missing_cb callback function -
introduced in the previous patch - to deinitialize the pipe. The
callback is called by PollEvents, which is synched with the pipe
lifecycle, avoiding the crash.

One interesting bonus is that file will be deleted only when
FileReader reached eof, because the missing notification is triggered
only after an eof notification. The original version could
deinitialize the pipe with handled file content.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>