hook-commands: add hook-commands() option to all source/destination drivers
This module adds a new plugin that is capable of injecting itself into both
source and destination drivers.  It makes it possible to execute external
programs when they are initialized or torn down.

Here's an example that opens an iptables port automatically as syslog-ng is
started/stopped.

Assumption: the LOGCHAIN chain is part of a larger ruleset that routes
traffic to it.  Whenever the syslog-ng created rule is there, packets can
flow, otherwise the port is closed.  This is a silly example, but I think
you get the idea.

source {
  network(transport(udp)
          hook-commands(setup("iptables -I LOGCHAIN 1 -p udp --dport 514 -j ACCEPT")
                        teardown("iptables -D LOGCHAIN 1")));
};

It might even be combined with something like salt for using more
declarative actions instead of simply coding add/remove transactions which
might break if syslog-ng dies.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>