cfg-block: only use __VARARGS__ if they are explicitly requested by the user
Blocks are very powerful, especially with __VARARGS__ being able to pass
even unknown arguments over to a specific processing element. However,
the default behavior of not validating arguments to blocks (and rather
adding them to __VARARGS__) means that a typo is very difficult to
recognize.

So this patch requires '...' at the end of the block argument list to
turn on the __VARARGS__ behavior, and even if enabled, unknown arguments
are reported to the debug log.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>