nvtable: deserialized handle invalid if fixed up
The deserialization process for nvtable has two stages.
The first one is to read the byte stream and assign it to a structure,
The second phase called `fixup`, which tries to fix the index tables (both static and dynamic).

The fixup process checks if the handles are valid, and current. If they are not, it replaces them with a new valid handle.
The current issue is that the fixup function created a temporary table for all of the index, and saved all of the updated handles in this new table.
If there was even one new handle, it swaped the two table.

But in fact if there was a handle that did not required update, it was not saved to the new table, thus that space left uninitialized.
Thus it could crash based on the memory trash.

Signed-off-by: kokan <peter.kokai@balabit.com>