Enforce use-dns-cache(no) when use-dns(no) is set
When using the DNS is disabled, disable the DNS cache too. We do this both for
the global and the per-source case too, by way of doing it in a place called by
both. As there are no other places where we set these, enforcing at these two
places is enough.

Whenever we enforce the above, we also issue a warning. Sadly, we do not have
access to any object that would allow us to emit the location of the setting, as
that information does not trickle down to where we do the enforcement.

Signed-off-by: Gergely Nagy <algernon@balabit.com>