plugin: reverting  03771dbba823483e0b285c1c4c8a60c5d16a6f57
The original code scanned the plugins (shared libraries) in a separate
process and sent the list of plugins to the master process, then
exited. This means the main process could selectively load the
available plugins. Unfortunately due to unknown reasons, syslog-ng
fails to start with this version. We are reverting the patch, and
later we might try to reintroduce this functionality when the root
cause of the sympthom is found.

Running valgrind with --run-libc-freeres=no would start syslog-ng, but
causes a lot of false positive results (even a simple printf appears
in a log). The problem seems to be connected with the fork() and
exit(), but the root cause is yet to be known.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>