elastic-v2: Force a flush on send if an out-of-band flush failed
Because `flush` can be - and is - called out of band, any error it may report
does not get back to syslog-ng when called this way. This in turn can result in
undeliverable messages being stuck in the bulk batch, until the retry mechanism
kicks in. And that can take a long time.

To lower the severity of the problem, a new flag is introduced, which is set by
`flush` when it fails. The next time `send` is called, if this flag is set, it
will force a `flush`, to trigger the retry mechanism properly. Once the `flush`
succeeds, the flag is cleared.

We rely on the retry mechanism provided by `LogThreadedDestDriver` to clear out
undeliverable messages from under us.

This is not the best fix, isn't even a proper fix, but it reduces the impact of
the problem considerably, thus addressing the main concerns of #1993.

Signed-off-by: Gergely Nagy <algernon@balabit.com>