mainloop-worker: fix thread_id allocation for more than 32 CPUs
We are storing the allocated thread ids as a bit array stored in a
guint64. However when trying to check single bits, we used an int
to shift into the various bit locations. ints are 32 bits, thus
when we shifted 1 << 32, we got 0, which is not correct.

This patch properly adds the LL suffix to those literal numbers, making
sure that we actually shift 64 bit values.

This fixes a crash on machines with more than 32 CPUs/cores.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>