configure.ac: Fix mixed linking
When using mixed linking, we want to link libevtlog dynamically, and without
using libtool. For this reason, we can't use the '.la' file, and we want to link
it before we tell the linker to "link the rest statically".

Additionally, we need to specify the libraries in a certain order, otherwise the
linker will not be able to resolve symbols. The diskq and syslog-ng-ctl
Makefiles were adjusted accordingly.

Fixes #2020.

Signed-off-by: Gergely Nagy <algernon@balabit.com>