cfg-parser: Report memory exhaustion errors during config parsing
Instead of simply failing the configuration parse process when the parser
encounters a memory exhaustion, report the fact more properly. To do this, we
store the result of `self->parse`, so that we can check for the OOM state too.
If we are in that state, report the error with a suggestion. We still narrow the
return value of `cfg_parser_parse` to a boolean, with the OOM case obviously
being a failure case.

Fixes #2030.

Signed-off-by: Gergely Nagy <algernon@balabit.com>

squash! cfg-parser: Report memory exhaustion errors during config parsing

Signed-off-by: Gergely Nagy <algernon@balabit.com>