cfg-lex.l: fix location tracking for multi-line block-refs
For tokens that span multiple lines (in our case this only happens when
a block-ref is parsed that spans multiple lines), we blew up the line
number counter, so whenever reporting locations or errors we got an
incorrect value.

This patch fixes that by shifting to the new line number relative to
last_line, instead of simply increasing the first_line.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>