logsource: fix a race in log_source_flow_control_suspend()
In the previous implementation of log_source_flow_control_suspend(),
it was possible to lose acknowledged places from the window.

Between setting the suspended_window_size and setting the original window
size to 0, it is possible that window_size is incremented by another
message consumer in a different thread.

This window leak is permanent, it can not be reversed without reloading
syslog-ng. Since the problem can occur multiple times in a multithreaded
environment, the source may also get stuck.

Signed-off-by: László Várady <laszlo.varady@balabit.com>
Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>