amqp: remove rabbitmq-c submodule and its support
Now submouldes are half baked, some dependency are added as submodule and some are not.
As the AMQP is not a core functionality, so it is not a requirement to compile syslog-ng.

Also this is the last commit, where the autotools are supported by rabbitmq-c, unless we include the cmake into the build (integrate with autotools) there is no way to update the submodule, thus no fix can came from upstream. That's really bad.

Signed-off-by: kokan <peter.kokai@balabit.com>