afsocket: fix regression: support TLS in legacy tcp driver
Caught by functest where TLS is configured by using the legacy
TCP driver. The problem was that Multitransport was constructed with
default plain TCP transport and TLS related tests stalled.

The fix: Construct TLS transport when legacy TCP is used with TLS.
When transport() is not used and we have TLS context it means that user
configured TLS with legacy TCP driver.

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>