logthrdstdrv: get rid of spurious log_queue_reset_parallel_push() calls
A lot of log_queue_reset_parallel_push() invocations were scattered
around the code, making it a lot more complicated. It seems that most
of those is not needed at all, but fortunately are harmless.

This patch sets up a clear strategy in cancelling the parallel push callback,
and also documenting the same in the _perform_work() function as a comment.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>