affile: KW_FILE token was in global namespace
The following grammar does not parsed successfully:

parser test_parser {
channel {
if (message('foo')) {
filter {message('foo2')};
parser { date-parser();};
destination { file("/tmp/out2" ); };
};
};

Because it said it does not expect the `file` in the destination, although it worked with every other destination (network,redis, ...).
As it turned out the KW_FILE was in a global namespace, and could be reached before the actual file plugin, but of course that was a not expected symbol.