cfg-args: cfg_args_contains in cfg_block
While validating VARARGS, we check all user-defined parameters are
listed in the block definition. All parameters are stored as key-value
pairs in a hashtable. In the original implementation we checked the
value of the specific key, and if it was null, then we knew the key
was not listed. Having NULL for value of the key was equivalent to
that the key was not listed.

In the next patch, I would like to change this and use the NULL value
to express that a block was listed, but without value (meaning it is a
required parameter), so this logic would break. To resolve this, I use
g_hash_table_lookup_extended to check if key is available in the
hashtable.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>