cfg-block: empty brackets in parameter means required
The goal of this patchset is that block definitions can express that a
parameter must be overridden by user.

For this, I needed to change the block definition. Originally for an
empty block (), the lexer return with an empty string: "". I modified
the block definition to return with NULL in this case.

With this, I can loop through the parameters in the block definition,
and when I see NULL as value, I check if the user used that parameter
in the block reference.

Some complex scl-s (like default-network-drivers) would profit if
there was a way to specify a default value that expands into
nothing. For that usage, I modified the block definition again: in
case a block blockname(param("")) (i.e param contains an empty
doublequote), then it will expand into nothing

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>