dbparser: do not add synthetic messages to context
Adding synthetic messages to end of the current context seems intuitive,
but its side effects are not intuitive at all.

For example:
Triggering a context close event with "$(context-length)" == "N", where
the user wants to include $(context-length) in the message looks really
weird.

grouping-by(
    key("${IMPORTANT}")
    trigger("$(context-length)" == "5")
    aggregate(
        value("MESSAGE" "We have $(context-length) important messages out of 5 within 10 seconds.")
    )
    timeout(10)
);

The output can be the following prior to this patch:

We have 6 important messages out of 5 within 10 seconds.

Signed-off-by: László Várady <laszlo.varady@balabit.com>