file-perms: Default to -1 for file and directory owners
Default to -1 for file and directory ownership, meaning no ownership changes
will be applied, unless explicitly specified. This becomes important when
running syslog-ng as a user (with appropriate capabilities granted), because one
would expect that - unless configured otherwise - it will use the running user
for the owner of files it creates.

Fixes #2002.

Signed-off-by: Gergely Nagy <algernon@balabit.com>