logmsg: log_msg_clear always creates new nv_table instead of clearing current
log_msg_clear is used when we want to reuse a log message, but
we don't need its contents (SDATA, tags, NVTable), just
its properties (log path related properties, ack, etc.).
It's used only in msg-format handlers (syslog, kmsg).

This function had a condition which decided if we are
allowed to modify the NVTable or we should create a new one
(protecting other users of the log message).
The condition was not perfect and it can result
in an assertion failure when nv_table_clear called (NVTable ref_cnt), as
in case of parsers it happened.

Instead of improving the condition, which could only be achieved
with a layering violation (checking the NVTable ref_cnt),
it was decided that we should always use a new NVTable
and conditionally unref the previous one if it is not used by anyone.

Removing this conditional nv_table_clear results
in no performance degradation in the fast path,
this is an error handling.

As nv_table_clear is not used anywhere else I have removed it.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>