logqueue-fifo: memory_usage - possible underflow
In the original code, the content of backlog is dropped earlier, then
the empty list is iterated through to update memory_usage
counter. This means when messages put back to backlog later again,
the size was subtracted twice, hence the overflow.

This patch changes first updates statistics, then drops the content of backlog.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>