logthrdestdrv: separate event registration and initialization
In the unit test program covering logthrdestdrv, we are operating in a
different context, e.g. the startup of the worker thread may not be
guaranteed.

To support this context, make sure that all ivykis events/tasks are
initialized during the constructor phase. This makes them usable, even
if the thread has not - yet - started up.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>