afinter: mark mode - no need to update watches
update_watches will be called in afinter_source_post later.

The problem with this version is update_watches reschedules
mark_timer, but in this code path, the timeout was not increased. It
is only increased when the mark message is actually added to the
queue (log_src_driver_queue_method). In older ivykis versions, this
caused infinite loop: the handler of the timer in the past is called
infinitely. With newer ivykis versions, the handler of the passed
timer is not called immediately, so syslog-ng could call the queue for
the message, which would increase the timeout, avoiding infinite loop.

https://github.com/buytenh/ivykis/commit/7190532aca902865ec8e578ac1c23d504f52a75a

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>