java: keep JVM alive until syslog-ng lives
Once a java is configured, and the JVM is loaded, and it is not possible to safely unload the JVM.
Because there cannot be two or more JVM in one process, the second time syslog-ng tries to create a JVM it wont succeeed.

The plan is to keep the JVM once created as long as syslog-ng is runnig.

Signed-off-by: Kokan <kokaipeter@gmail.com>