system-source: removed container detection from system source, because changes in kernel code made it unreliable.
The function was used to avoid scenarios where the log management
on the host system and the containerised solution are collecting
the kernel logs simultaneously.
(In a lucky situtation this will only lead to "double collection",
but there might be cases where one process can "starve" the other.)

To keep the functionality (and make it user controllable) introduced
the "exclude-kmsg" option to system source. If "exclude-kmsg(yes)"
provided, syslog-ng will omit kernel logs on platforms where they are
available separatedly. (Currently Linux without systemd, and FreeBSD.)