afsocket: Add ref-counted TLSVerifier struct
When a TLS session is initiated, the afsocket module passes a
reference to the AFInetDestDriver instance to the TLSSession
structure to use as verification data. During config reload,
afsocket_dd_save_connection() may be called to store the driver's
LogWriter so that it will survive the config reload, and the
driver's 'writer' field is set to NULL. The driver is freed
without freeing the writer, thus the TLSSession still exists and
has a dangling pointer to the driver. The afsocket module may
then try to verify the session using the invalid data, which
results in a crash.

This patch adds a new struct, TLSVerifier, which stores the
verification data for a TLSSession, and a pointer to a function
for destroying the verification data. It also contains a
reference counter that is incremented when it is passed to a
TLSSession, and decremented when the TLSSession is destroyed.
This ensures that the TLSVerifier is not destroyed until it is
not referenced by any more sessions.

It is not necessary to use the entire driver as verification data
so the required info (TLSContext and hostname) is placed in a new
struct, AFInetDestDriverTLSVerifyData, which is used by the
TLSVerifier as verification data.

Signed-off-by: Bernie Harris <bernie.harris@alliedtelesis.co.nz>