loggen: fix file parser to support both log format
loggen was able to parse BSD (RFC3164) formated loglines
read from file. With this patch loggen is able to parse both RFC3164
and RFC5424 (syslog format).

It also fixes the possible buffer overrun in the old parse_line
function.

The deafult value of --skip-token has been changed to 0
as this is more intuitive as the old value 3. You have to use
the --skip-token option if and only if you want to skip the first N
token of the loglines. Otherwise you shouldn't care this option.

Signed-off-by: norberttakacs <norbert.takacs@balabit.com>