scl/ewmm: rename syslog-ng() source to ewmm()
syslog-ng is just an alias for the ewmm destination there's no syslog-ng()
source and shouldn't be one, even if it sounds asymmetrical.

Rationale:
  - on the source side we want ewmm to be processed on the same channel as
    everything else is coming in.
  - we already created default-network-drivers() for exactly this use-case
  - the ewmm() source below is a pretty limited use-case that is probably
    only good enough for testing.
  - with this, we have symmetry: ewmm() source and destination. And we
    also reserved the name "syslog-ng()" to send messages to an actual
    syslog-ng instance.

if someone needs a different use-case, the recommended way is to construct
the source plus parsing bits at the configuration file level, as that
use-case is not generic enough.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>