basicfuncs: add url-encode alias for urlencode
The convention is to add dashes to separate words in the name of the
template function (to match the convention in functional languages), and
albeit with urlencode wasn't that long, I didn't want to add base64encode
all as one word.

This patch adds an alias to urlencode, so that it can be used as url-encode.

With that, base64-encode would be added with a dash.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>