gprocess: obtain capability from text at runtime
The CAP_ macros come from the build machine kernel headers, the list/value is binded to the builder machine's values.
This function asks from the libcap the value in runtime. This shold be more portable.
Also this version check if the libcap supports the capability and also the kernel if supports it.

This is not optimal, because it does all the time, caching could be considerd.

Signed-off-by: Kokan <kokaipeter@gmail.com>