hdfs: disable archive when append is used
Archiving and append are 2 conflicting use cases.
Archiving is for moving closed file to a dir declaring that
processing of those files are finished.
If we would allow appending an archived file it would not really be archived.

Originally I suspected a bug when the two options used simultaneously,
however the rename() call from hdfs lib prevents overwriting an existing file.
This patch fixes the current issue of archiving a file once, then using
the original file indefinitely.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>