http: add batch delimiter option
This option allows one to specify the delimiter to be used between
elements of the batch. Previously it was assumed to be part of
the template (in case body() was specified), or was simply assumed to be
'\n' (in case it wasn't).

Fortunately the previous behavior wasn't released in an official release,
so we are still in time to fix this up.

The use-case is that the delimiter should not be generated neither before
the first or after the last element, as it causes syntax errors if
we generate a json.

Here's an example config:

                http(
                        url("http://localhost:8080/")
                        headers("Content-Type: application/json")
                        body-prefix("{\"messages\":\n[")
                        body("$(format-json --scope nv-pairs)")
                        delimiter(",\n")
                        body-suffix("]}\n")
                        flush-lines(100)
                );

This generates a properly formatted json array.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>