logthrdestdrv: introduce flush_lines member
Previously the flush_lines() option and the associated member was
defined in derived classes, however, there's a flush-lines() global option
as well, so deriving that value from GlobalConfig would have to be
implemented in each derived class.

This patch changes that and introduces this member at the
LogThreadedDestDriver level and then relies on the derived class to
actually use it.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>