hdfs: time-reap option for files
The new *time-reap* option mimics how the file destination version /should/ work.
When the value is configured and larger than 0, the default is 0 that means disabled.
If any of the file is not written in that amount of seconds, that file is closed
and/or archived - if archive is configured.

Signed-off-by: Kokan <kokaipeter@gmail.com>