logthrsource: move iv_init() and iv_deinit() into LogThreadedSourceDriver
The implementation of mark messages (mark-freq()) require all syslog-ng
source threads to be initialized with iv_init().

This is because mark timers are based on ivykis, and they are updated from
the source thread, using iv_now (a cached time object).

The mentioned timer update can be found in ml_batched_timer_postpone() and
in afinter_postpone_mark().

Signed-off-by: László Várady <laszlo.varady@balabit.com>