modules: add examples
The purpose of this module is to provide example plugins for developers.

Each plugin contains a short description in its README file.

Example plugins are categorized into folders based on their type (sources,
destinations, ...).
Every plugin is a separate static library, but they will NOT be installed.
The `examples` module is linked against them instead.
After compilation, one single `libexamples.so` module is created and installed.

This commit adds the first example module, the msg-generator source.

Signed-off-by: László Várady <laszlo.varady@balabit.com>