examples: add threaded random-generator() source
This example source module generates `bytes()` random bytes every `freq()`
seconds. The output is represented as a hexadecimal string in a generated
log message.

`type()` can be `random` or `urandom`, the default is `random`.

The plugin is implemented on the top of `LogThreadedSourceDriver`.
It provides a separate thread, so blocking function calls are allowed.

Signed-off-by: László Várady <laszlo.varady@balabit.com>