lib/gsockaddr.c: modify the unix salen calculation
It is not required on all platforms to add the size of the
terminating null character (+1) to the value of the addrlen
parameter during the bind system call. Most platforms has a
built in MACRO to calculate the proper value, called SUN_LEN.
SUN_LEN is not a POSIX standard, so we use the formula from
the Linux man page if it is not available.

Signed-off-by: Laszlo Szemere <laszlo.szemere@balabit.com>