value-pairs: add support for --shift-levels
The new transformation would skip the specified number of dot-separated
tokens from the beginning of the keys. This way it is possible
to remove a json encapsulation from a set of keys.

Usage:

$(format-json --key .SDATA.* --shift-levels 2)

This would drop the .SDATA. prefix.

The numbering basically denotes the number of dots to remove, and
the initial dot counts.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>