value-pairs: fix type hint parsing
Type hints are embedded in the value part of the syntax, e.g.

$(format-json foo=int($WHATEVER))

Here, "int" is the typehint and we'll find the actual value enclosed
in parentheses. The problem happened if the value itself contains
a closing paren, as the type hint parsing does not really count
opening and closing parentheses, it was simply looking for the closing one
from the start.

This patch basically reverses the lookup, e.g. we are looking for the
closing paren from the end.