stats-cluster: add autoreset support
autoreset: when a StatsClusterInstance has no further users
(use_count==0) then all the autoreset-enabled counters are set to 0.

motivation: make it possible to share counters between multiple threads

Note that the autoreset mask is added also to StatsCounterGroupInit, and
when the group_init method is called, it will set the mask for the
StatsCounterGroup initialized by the group_init method.
Why? Because it should work well for all StatsClusters, even for
SingleClusters.

( some additional note:
* StatsCluster has-a StatsCounterGroup, which stores the counters
* when a StatsCounterItem is registered as first from a StatsCluster,
the StatsCluster does not exists, it is created during registration:
this is why the StatsCounterGroupInit is part of the StatsClusterKey and
this is why the autoreset mask is store in the StatsCounterGroupInit
* if we don't want to support autoreset counters for SingleClusters,
then it is enough to store the autoreset mask in the StatsCounterGroup,
and can be set by the group_init )

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>