logqueue: use log_queue_length() instead of cached value when registering ctrs
The cached value for queued_messages cannot be use after a reload.

When a LogQueue instance is found the very first thing what are doing
with that is unref (in case of diskq).

On the other hand, we need to track queued_messages to ensure that the
StatsCounter is 0 after a deinit (without the autoreset feature).

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>