logthrdestdrv: allow the thread startup to fail
This patch introduces the concept of thread startup failing, that can be
used to propagate thread specific startup-error conditions back to driver
initialization. An example use-case is libcurl initialization that can
fail.

Should such a failure happen, driver initialization is reported to be
failing, thus syslog-ng will refuse to start.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>