host-resolve: specify ai_hints explicitly
Pass AI_V4MAPPED | AI_ADDRCONFIG as hints to getaddrinfo() explicitly, which
allows ip-protocol(6) destinations to fall back to ipv4, should the
ipv6 stack not be available or if the server does not have an ipv6 address.

This would allow the user to use either ipv4 or ipv6, depending on the
ipv6 connectivity and the contents of DNS.

See getaddrinfo() manual page for more information.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>