afsql: use the LogThreadedDestDriver based infrastructure
This patch removes most of the private infrastructure that existed in the
afsql driver and relies LogThreadedDestDriver for the same functionality.
The previous patch already changed the base class without connecting in its
methods, this patch now calls the appropriate methods and also compiles.

Subsequent patches would further clean the code up, the goal here was to make it compile.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>