gprocess: enable_caps introduced
The *caps* variable were used to determine if the *caps* support was enabled or not.
In reality there are 3 possile state that we should code:
* default caps
* no caps
* caps from user

It is possible to achieve this with only the *caps* by initializing it with default caps,
and let the user change the value after that step. If it is not possible to initialize before the user supply
different caps value, the above solution won't fly anymore.

Thus this introduces the *enable_caps* variable to represent all of the states.

(This is needed so the user supplied *debg_flag* could be accessed. See next patches.)

Signed-off-by: Kokan <kokaipeter@gmail.com>