nondumpable: fix possible binary portability issue
If someone builds syslog-ng on a platform where MADV_DONTDUMP is
available, then the created binary cannot be used on systems where it
it is not available. So we need to detect the availability at runtime
too.

To be specific: centos6 backported this feature into their libc version with
https://access.redhat.com/errata/RHBA-2013:0279, even though from the
manpage, this option should be only available from kernel 3.4. The
resulting binary cannot be used on ubuntu:precise, as it uses 3.2
kernel.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>